minor comments
