+++ pkg/async/notifications/implementations/aws_emailer.go
@@ -11,25 +11,8 @@ import (
 	"github.com/flyteorg/flyteidl/gen/pb-go/flyteidl/admin"
 	"github.com/flyteorg/flytestdlib/logger"
 	"github.com/flyteorg/flytestdlib/promutils"
-	"github.com/prometheus/client_golang/prometheus"
 	"google.golang.org/grpc/codes"
 )
-
-type emailMetrics struct {
-	Scope       promutils.Scope
-	SendSuccess prometheus.Counter
-	SendError   prometheus.Counter
-	SendTotal   prometheus.Counter
-}
-
-func newEmailMetrics(scope promutils.Scope) emailMetrics {
-	return emailMetrics{
-		Scope:       scope,
-		SendSuccess: scope.MustNewCounter("send_success", "Number of successful emails sent via Emailer."),
-		SendError:   scope.MustNewCounter("send_error", "Number of errors when sending email via Emailer"),
-		SendTotal:   scope.MustNewCounter("send_total", "Total number of emails attempted to be sent"),
-	}
-}
 
 type AwsEmailer struct {
 	config        runtimeInterfaces.NotificationsConfig
