+++ pkg/async/notifications/implementations/sendgrid_emailer.go
@@ -22,6 +22,7 @@ type SendgridEmailer struct {
 func getEmailAddresses(addresses []string) []*mail.Email {
 	sendgridAddresses := make([]*mail.Email, len(addresses))
 	for idx, email := range addresses {
+		// No name needed
 		sendgridAddresses[idx] = mail.NewEmail("", email)
 	}
 	return sendgridAddresses
@@ -29,7 +30,6 @@ func getEmailAddresses(addresses []string) []*mail.Email {
 
 func getSendgridEmail(adminEmail admin.EmailMessage) *mail.SGMailV3 {
 	m := mail.NewV3Mail()
-
 	from := mail.NewEmail("Flyte Notifications", adminEmail.SenderEmail)
 	content := mail.NewContent("text/html", adminEmail.Body)
 	m.SetFrom(from)
